Allow keeping the dropdown open when items are clicked.
For some scenarios, it is very beneficial if the dropdown menu is not
closed when items in the dropdown is clicked. An example is a button
that shows a dropdown with a form inside, e.g. login controls.

This commit adds a "keep-on-menu-click" attribute that can be set to
true when you want the menu to stay open on internal clicks.